moved infinte files back to resonantcontroller
