+++ vitis/software/Baremetal/src/uz/uz_ResonantController/Resonant_Controller_ert_rtw/Resonant_Controller.c
@@ -21,8 +21,10 @@
 
 #include "Resonant_Controller.h"
 #include <math.h>
-#include "../../../Codegen/uz_codegen_shared/rt_nonfinite.h"
-#include "../../../Codegen/uz_codegen_shared/rtwtypes.h"
+//#include "../../../Codegen/uz_codegen_shared/rt_nonfinite.h"
+//#include "../../../Codegen/uz_codegen_shared/rtwtypes.h"
+#include "rt_nonfinite.h"
+#include "rtwtypes.h"
 #include <string.h>
 
 /* Model step function */
