+++ vitis/software/Baremetal/src/uz/uz_ResonantController/Resonant_Controller_ert_rtw/rtGetNaN.c
